Add files via upload
The .rmd file gave me an error where it was describing the hip circumference variable as null in the Brozek fit, but the code was still able to run. In the case that the data does not show on the .Rmd file, I have included a text file with the code in it.